# server/src/docs/api_swagger.yml
swagger: "2.0"
info:
  description: "This is the documentation of SendIT API hosted on Heroku"
  version: "1.0.0"
  title: "SendIT"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "oesukam@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "andela-sendit-api.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWT_Token_Admin:
    type: apiKey
    in: header
    name: Authorization
  JWT_Token_User:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /users/{userId}:
    put:
      summary: "Update user's names, birth date, province, district and city"
      security:
        - JWT_Token_User: []
      tags:
        - "users"
      operationId: updateUser
      parameters: 
        - name: userId
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: return the basic user information
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                $ref: "#/definitions/User"
    get:
      summary: "Fetch basic user's information"
      security:
        - JWT_Token_User: []
      tags:
        - "users"
      operationId: getUser
      parameters: 
        - name: userId
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: return the basic user information
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                $ref: "#/definitions/User"
  /users/{userId}/parcels:
    get:
      summary: "Fetch a particular user's parcels"
      security:
        - JWT_Token_User: []
      tags:
        - "users"
      operationId: getUserParcels
      parameters: 
        - name: userId
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: return a list of user's parcels
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                type: array
                items:
                  $ref: "#/definitions/Parcel"
...
definitions:
  User:
    type: object
    properties:
      id:
        type: string
        format: uuid
      email:
        type: string
        format: email
      first_name:
        type: string
      last_name:
        type: string
      province:
        type: string
      district:
        type: string
      gender:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      updated_at:
        type: string
    required: 
      - id
      - email
      - first_name
      - last_name
      - province
      - district
      - gender
      - created_at
      - updated_at
...
  Parcel:
    type: object
    properties:
      id:
        type: string
        format: uuid
      user_id:
        type: string
        format: uuid
      from_province:
        type: string
      from_district:
        type: string
      to_province:
        type: string
      to_district:
        type: string
      receiver_names:
        type: string
      receiver_phone:
        type: string
      receiver_address:
        type: string
      weight:
        type: number
      price:
        type: number
    required: 
      - id
      - user_id
      - from_province
      - from_district
      - to_province
      - to_district
      - receiver_names
      - receiver_phone
      - receiver_addres
      - weight
      - price
...